v2/requirements.txt
einops
gymnasium>=1.2.3
mediapy
numpy
pillow
pyboy>=2.7.0
stable_baselines3>=2.7.1
tensorboard
websockets
